feat(network): define config for rate limit threshold (#1284)

+++ network/ratelimit_test.go
@@ -8,16 +8,16 @@ import (
 )
 
 func TestRateLimit(t *testing.T) {
-	threshold := uint8(5)
+	threshold := 5
 	window := 100 * time.Millisecond
 	r := newRateLimit(threshold, window)
 
 	t.Run("InitialState", func(t *testing.T) {
-		assert.Equal(t, uint8(0), r.counter)
+		assert.Equal(t, 0, r.counter)
 	})
 
 	t.Run("IncrementWithinThreshold", func(t *testing.T) {
-		for i := uint8(0); i < threshold; i++ {
+		for i := 0; i < threshold; i++ {
 			assert.True(t, r.increment())
 		}
 		assert.Equal(t, threshold, r.counter)
@@ -30,14 +30,14 @@ func TestRateLimit(t *testing.T) {
 	t.Run("ResetAfterWindow", func(t *testing.T) {
 		time.Sleep(window + 10*time.Millisecond)
 		assert.True(t, r.increment())
-		assert.Equal(t, uint8(1), r.counter)
+		assert.Equal(t, 1, r.counter)
 	})
 
 	t.Run("ResetMethod", func(t *testing.T) {
 		r.reset()
-		assert.Equal(t, uint8(0), r.counter)
+		assert.Equal(t, 0, r.counter)
 		assert.True(t, r.increment())
-		assert.Equal(t, uint8(1), r.counter)
+		assert.Equal(t, 1, r.counter)
 	})
 
 	t.Run("DiffMethod", func(t *testing.T) {
